Store the original role attributes as a shallow copy
This was the original intention since it leaves EquivClasses for the
announcing class but strips it from the role attributes in the
equivalent classes themselves.

The array ref value of EquivClasses is still modified by RegisterRole to
add RT::System, but that's OK since it is simply an implicit equivalent
class.  It should be introspectable later rather than hidden.