Introduce the ACLOnly and ACLOnlyInEquiv flags to role metadata
These are advisory flags indicating when a role is only used for
granting rights and generally shouldn't contain any members.  For
example, the Owner and Requestor ticket roles should not contain members
at the queue or system level.  These flags are useful in the UI when
generating a list of selectable roles to display for objects.