Restrict SUBCLAUSE lexically during parsing, and switch to simple Limit
Now that Limit functions as every other DBIx::SearchBuilder class, it is
no longer necessary to call SUPER::Limit.  Additionally, the use of the
confusing _SQLLimit method can be avoided by lexically setting the
default subclause during parsing in FromSQL.  This obviates the need to
remember the difference between the two methods.

Deprecate the old _SQLLimit for future removal.