Check the return value of RT::URI->FromURI instead of ->Resolver
FromURI always sets a resolver, even on failure.  Error checking which
only looks at ->Resolver is broken and checking ->Scheme on top of that
is just patching over the bug and relying on the internals of
RT::URI::base.  FromURI returns a useful success or failure value; just
use that.