Tests: Enable any plugins listed in RT_TEST_PLUGINS
RT_TEST_PLUGINS should be a space-separated value containing a list of
plugin names (the same names that go in @Plugins).

The root directory for each plugin may be set via an environment
variable as well, in case the plugin isn't installed in the normal
location.  These environment names take the form of:

    RT_TEST_PLUGIN_<plugin name>_ROOT

and the value should be an absolute path.  The <plugin name> segment may
preserve the double colon package separators, omit them entirely, or
replace them with underscores (_) or hyphens (-).  For example, all of
the following are legal for a plugin named "RT::Extension::Example":

    RT_TEST_PLUGIN_RTExtensionExample_ROOT
    RT_TEST_PLUGIN_RT::Extension::Example_ROOT
    RT_TEST_PLUGIN_RT_Extension_Example_ROOT
    RT_TEST_PLUGIN_RT-Extension-Example_ROOT

The overriding of plugin roots (but not enabled plugins) was previously
available via similarly named CHIMPS_<plugin name>_ROOT variables (to
support the Test::Chimps framework, now abandoned).

The upshot of these simple changes is the ability to run all or parts of
RT's core test suite with one or more plugins loaded, thus testing the
plugins' effect (or really, lack thereof) on core operations.