Restore QueueWatcher conditions to TicketSQL and the collection API
The class on which we want to look for roles was lost during the
refactoring of role limits and joins in commit 4b15643.  This commit
restores the ability to tell RoleLimit that you're interested in roles
on containing classes like RT::Queue.  It also allows the same logic to
happen with arbitrary other record types using roles, removing the
hardcoded "Queue or id" column choice.