Allow ListMenu to show 2 levels of menus
For Searches, it's kind of tedious to only see Tickets/Articles/Users,
really we want to maintain the older links to New Search and Simple
Search, etc.  This is less needed in the /Admin/ tabs where you are
expected to navigate to another Admin/Foo/index.html, but in Search
where you jump right in to Build.html or Simple.html, it helps to see
all the options.