Scrub HTML from the spreadsheet view of ticket searches
This prevents the web display parts of column map from leaking into the
spreadsheet, such as through custom fields.  However, multiple value CFs
are now output all run together instead of separated by "<br>" elements.
Whether this is an improvement or not depends on your point of view, but
it will be remedied to something saner in a following commit.

Any HTML entities are also decoded so that no HTML should be left.