Ensure that ->_Accessible builds attributes when called as a class method
Prior to 5671e60, ->_Init was guaranteed to have been called prior to
->_Accessible, as the former is called during ->new, and the latter
required an instance.  When calling it as a class method, that invariant
does not hold, and autovivification causes an empty hash to be used.
This, in turn, causes all column fetches on records of the class to fail
with "___ unimplemented".

Fill the attributes cache using _BuildTableAttributes, if it is not
already.