Rejigger Tickets and Articles under a Search tab
I think this is clearer than remembering that 'Tickets' is where you
click to search.  It does bury the Prev/Next links for navigating a
search deeper, but we may want to make those more accessible instead of
burying them up here (such as in the floating actions).

This also means that Search wants a top level tab like /Admin/ has.
Showing more than one level of menus resulted in the css top margin
looking really gappy and raggedy so that's been snugged up a bit.

Very few tests failed as a result of this change.
  /Articles/ matches the search instead of the top level articles UI
  The id of the 'New Search' link changed
  Clicking on Search got you the Search listing, rather than the
    Articles Search.