Encapsulate the role group class into a method
This enables overriding of the default of blessed($self->NewItem).
NewItem may return a subclass of RT::Ticket, which doesn't match the
database values and confuses the role limits.  An example is
RT::Reports::Tickets which uses a NewItem of RT::Reports::Tickets::Entry
but cares about role groups on RT::Ticket.

The method is provided by the Roles role, so most consumers need not
worry about this implementation detail unless they're doing something
wonky.