Linkify filenames in POD (F<...>) which point to documentation
Now POD expressions such as F<docs/full_text_indexing.pod> will generate
an HTML link to the correct doc with the filename as the link text.  The
following are examples that now produce links:

    F<docs/full_text_indexing.pod>
    F<lib/RT/Date.pm>
    F<RT/Date.pm>
    F<rt-email-group-admin>
    F<RT_Config.pm>

This behaviour is particularly useful when referring to documentation
that isn't a Perl package file (.pm).