Use a pure Plack and HTTP::Message solution for internal PSGI requests
Replaces LWP::Protocol::PSGI and the complexity wrapped up in that
module with a much more straightforward approach.  Since we're writing
new code — not pulling the PSGI-wool over the eyes of otherwise
unsuspecting code — we don't need the cloak of LWP.

A nice side benefit is that without LWP::Protocol::PSGI, there are no
warnings when LWP::Protocol::https isn't installed.  Thus, we won't need
to either require it for the squishing functionality or suppress the
warning with ugly warning handlers.