User and custom field autocompletion for Self Service
Tiny shims simply redispatch to the normal autocomplete endpoints,
letting us easily control which endpoints are available.

The Users endpoint is safe for unprivileged access since it is designed
with that in mind and protected via the $AllowUserAutocompleteForUnprivileged
option.  This commit resolves [rt3 #16946].

The CustomFieldValues endpoint is also safe for unprivileged as it
strictly follows granted ACLs and context objects.

To simply the endpoint decision in JS, I added WebHomePath to the
javascript config data, mimicking the same value available to column
maps since 159e206.