Instantiate new sessions on logout as well as deleting the old one
Although the stored session is deleted by tied(%session)->delete, the
tied object and all data still remains present in memory.  When %session
goes out of scope at the end of the request, the Apache::Session DESTROY
block attempts to save the session one last time.  On database session
stores (MySQL and Pg), the update silently fails because the row has
been deleted (and no error checking is done).  When using
Apache::Session::File, however, session updates autovivify the file,
essentially reverting the deletion.  The autovivified session contained
an empty current user, since the CU is replaced after the delete, but
all other user-specific session data still existed and was in the
session of the unauthenticated client.  This lead to incorrect behaviour
in extensions outside of core RT.

The old session id in the client's cookie is now immediately replaced
with a new session id.  Previously it was replaced it on the very next
request (meta-refresh to the login page) when the old id could not be
loaded.  Under Apache::Session::File, the old session was not replaced
at all until a successful login.