Instantiate a new session if the session doesn't match the ID we loaded it by
The replaced code was a no-op since the refactor in 64f65d8.  Before the
64f65d8, deleting from the parsed cookie data was enough to set a new
cookie.  The new cookie contained the found session id, regardless of if
it matched the loaded session id.

A safer, more conservative approach is to abandon the old session and
force a new one if the session cookie value loads a session with a
different id.  This should never happen but indicates corrupt session
data or malicious activity if it does.  There is no harm in the extra
protection.