Callbacks for changing the link destinations in the queue summaries
While likely to be used only rarely, these callbacks are very useful for
extensions which provide a richer "queue dashboard" than the vanilla
search results normally linked to by the queue summary portlets.
Without the callbacks, such extensions are forced to overlay the entire
files just to change some URLs.

The duplication between the two types of queue summaries is unfortunate,
but large scale refactoring is out of scope for the simple goal of this
commit.  A future refactoring isn't hampered any by the additional
duplication.