Revert "Switch from divs/spans to tables"
This reverts commit cb3b5a57b1e887fa28e648a8d35b7a0c9b06c86f.

Tables are annoying and hard to style well with CSS.  We lose
flexibility going forward.  The underlying reason for the poor
formatting noted in the reverted commit message is because the class
name was mistakenly changed in the refactor and the relevant CSS no
longer applied.  The class name will be fixed in the following commit,
and all shall be well.