HTML escape any user field titles provided by the configured format
This eases the use of &, <, >, etc in a format-supplied title (i.e via
/TITLE:Foo).  In contrast, ColumnMap supplied titles are allowed to use
HTML since they have another mechanism for proper escaping and expect to
be able to output HTML.

The title isn't scrubbed of disallowed HTML because it is only
changeable by the RT server admin.  Before this commit, scrubbing the
HTML was also moot since the Mason interpolation didn't turn off the
default HTML escaping.