AsString on Tickets and Articles now includes the Name
This allows us to remove the ugly code in ShowLink and make it ugly in
here instead.

This means that Formats which say __RefersTo__ or __HasMember__ will
now say "ticket #42: Ticket Subject" rather than the current state of
the world where that renders as "ticket #42" which is not helpful.

The code in rt.pm's AsString is ugly because it assumes that you might
use this on some non-RT::Ticket object.  That never really happens, but
all of this code pretends it might happen so I kept up the illusion.