Generalize UserInfo's use of ColumnMap to display a single record object
Displaying other individual record types is now very simple.  By
default, the passed in Format is untrusted and will be scrubbed of
disallowed HTML.  If the calling code knows the format is safe, perhaps
because it can only be from a config file, TrustFormat => 1 may be
passed.  This flag is present for use by UserInfo.