An invalid "results=" parameter is not a CSRF threat
Since results keys are removed from the session once they're used,
refreshing a page with a results= parameter would throw up the CSRF
interstitial.  That behaviour is almost as bad as a browser asking if
you want to resubmit the POST data on refresh, which is the reason for
the MaybeRedirectForResults() framework in the first place.