Exclude wasn't working with more than one record
This was generating (id != 7 OR id != 9) which gleefully returns both 7
and 9. Switch over to an ENTRYAGGREGATOR of AND. This is exposed in the
Membership page of Groups. Groups with one user had that user excluded
from the member autocomplete, but as soon as you have two members that
stopped working (same for groups).