return back documentation of DontSearchFileAttachments option
It was deleted in d789bd3986a4ad63a71bd2450c9aad0c7763a7de (rt-4.0.1rc1)
with usage in sbin/rt-setup-fulltext-index, but it is primary use
case is in TicketSQL to filter out named attachments while searching
content.

Mentioned commit replaces use of the option with command line argument.
It was the right call to separate FTS indexing from this option, so
we don't restore usage in the script.