Ensure that the error message is logged with newlines
die() is verridden somewhere in the stack to s/\n/\\n/g; this makes the
error message much less readable.

Use a direct call to $RT::Logger to ensure that newlines are preserved,
which also allows us to set the log level explicitly.  Not exiting via
die() requires that we handle providing the HTTP response by hand, which
also provides the opportunity to explicitly call out the
misconfiguration.