use RT::Date to parse cf date values in tickets search
to distinguish if user inputs date or datetime, here is the rule:

1. if cf type is Date, it's a date.

2. if cf type is DateTime, unless there is "hh:mm:ss" or "midnight", as
long as the parsed time is at midnight(00:00:00), it's a date.

we treat midnight datetimes as dates sometimes specially is because if user
searches a DateTime cf with a value like '=' '2013-03-21', we assume he means to
get all the tickets with the cf values in the range of the whole day.