Remove the local-install Makefile target
The local-install target was added in 578192f, but there was never a
shipped set of local/{etc,html,po} directories to install from (as far
back as we can reasonably go in svn history).  Thus the target always
errored out, although the errors were intentionally ignored.  Creation
of the local/ hierarchy in /opt/rt4 is handled by the "dirs" target.

The best guess is that this supported bundling of customizations when
building custom tarballs.  These days we have installable extensions.