Use ShowUser in ColumnMap, to link to user summaries
This alters formats to use the user's preferred display format, instead
of hard-coding the user's email address (for AdminCcs, Ccs and
Requestors) or the users's name (for Owners).  This makes user display
generally more consistent, although in some use cases the RealName of
users may obscure which Name or EmailAddress is involved.