Add a user format between concise and verbose based on Privileged
The "concise" format, when originally introduced in 07ec5dc, contained
similar logic which depended on the state of ->Privileged.  However,
this was removed in 20d879f, after it appeared to incur a notable
performance penalty.

However, upon closer analysis, much of the penalty was incurred by the
need to render the names of several hundred users for the ShowUser
dropdown -- a configuration which causes performance penalties
regardless.

At a possible minor performance cost, use the Privileged bit to control
how users are displayed.