Only display privileged users in Owner drop-downs
Configurations in which unprivileged users are expected to own tickets
are vanishingly small or non-existant; however, it is a common
misconfiguration to grant Everyone the OwnTicket right.

Limit entries in the Owner dropdown to only privileged users.
Configurations which require unprivileged users be able to own tickets
may use the autocompleter.