Load a Nobody user, not the Nobody current user
RT->Nobody returns a RT::CurrentUser, whose current_user is Nobody.
This object often cannot have its custom fields inspected by a custom
->Format -- even with f18f7ef -- because the custom fields might well
not be visible because they are loaded as the Nobody user itself.

Load a RT::User object as the current user, rather than using
RT->Nobody, which loads it as a user with no permissions.