override _Handle in RT::SearchBuilder
Like in RT::Record - just return $RT::Handle.

This avoids storing $RT::Handle in $sb->{'DBIxHandle'},
so it is never serialized in $session{'tickets'}. Such
serialization may lead to disconnection of active handle.
It happens when $DBIx::SearchBuilder::Handle::PrevHandle
is set to active handle and $session{'tickets'} gets
destroyed. It's hard to get triggered in a test.