Skip ckeditor files for license tags, not just fckeditor files
19efd7e added rules to check for proper license tagging, pulling its
regular expression of paths to skip from the license tagger.  The
latter, however, had not been updated for the FCKeditor -> ckeditor
change of 4d1ccc6, leading to false positives in the recent ckeditor
upgrade.

Skip all paths that contain ckeditor, not fckeditor, as the latter is no
longer in use.