Valid HTML for emailed dashboards
Commit b0ec8b3 added the opening <div> to satisfy the assumptions of
/Elements/Footer, but /Elements/Footer is not used when emailing the
dashboard due to RT::Dashboard::Mailer disabling Mason autohandlers.
This meant dashboard emails ended with the opening <div> and never
closed <body> or <html>.

Additionally the Footer and PageLayout *are* used when viewing
dashboards in the browser, so the extra opening <div> forced the footer
and remaining page content as a child of the "#body" div instead of a
sibling.

Fix this comedy of errors by properly closing the HTML document when in
email (with no footer) and letting the Footer close PageLayout's opened
"#body" div as expected when in the browser.