Position the page menu in aileron to avoid overlapping the inner shadow
The extra 0.05em accounts for the border pixels in the elements visually
above the page menu.

Unfortunately, the #page-navigation div is outside of the #body, which
means we can't take the easy route and position relatively up into the
corner.  We're left with approximating the distance from the top of the
viewport to the top of #body in CSS.  This distance is a combination of
element heights, paddings, and margins specified in ems (or 0.9ems for
every 1em when font-size: 0.9em) plus borders specified in pixels.

Resolves [rt3 #16806].