Change _CurrentUserHasRightToSetOwner to public CurrentUserCanSetOwner
Make public and rename to CurrentUserCanSetOwner to make it
consistent with other CurrentUserCan method names.

Meld _IsProposedOwnerChangeValid into CurrentUserCanSetOwner to
handle all of the logic in one place and provide a single
response from CurrentUserCanSetOwner. This simplifies SetOwner
and makes CurrentUserCanSetOwner suitable for calling to test
for owner change permissions in the context of the current
ticket.