Due to a git bug, be explicit about ignored directories
Git between 1.8.1.1 and 1.8.1.6 saw devel/third-party and would only
match against a *file* named devel/third-party, it required you to say
devel/third-party/ if you wanted to match the directory.

This was fixed in efa5f8254 in git.git, but since release engineers
might have one of the broken versions of git, we'll add a trailing
slash.