Force internal RT values for ticket Type to lower-case
The three values for Type that RT uses internally are 'ticket',
'approval', or 'reminder'; for the most part, these are not exposed to
the user.  The exception to this is when generating approvals, as the
CreateTickets action requires the user to specify the Type of the ticket
created.  Improper casing on the "approval" type is a common, and
frustrating, mistake.

Be liberal in what case RT accepts for its core types, but consistently
use lower-case in those values when they are stored.  As external
systems may rely on case-sensitivity of non-core values for Type,
preserve the case sensitivity of those values.