Canonicalize on lower-case for statuses, which are now case-insensitive
Prior to the merge of lifecycles (in 2ad74ba), statuses were fully
case-sensitive.  This meant that attempts to set the status to 'OPEN'
would fail, as the case did not match the supported 'open' status.
Lifecycles made checks for _validation_ of statuses case-insensitive,
but did not spread that property to many other parts of the system --
including rights, transitions, and searching.

Canonicalize all statuses to lower-case in the database, and make the
remaining entry points into the API lower-case as well.  This also
requires checking that all statuses described in lifecycles are
correctly in lower-case.  While using perl 5.16's "fc" operator would be
more correct in this case, this is not possible on all versions of Perl
that RT currently runs on.