warn on case sensitive searches
Because of historical reasons, RT's searches were marked as case
sensitive in RT::SearchBuilder::Limit, although callers often expected
them to be case insensitive.

Mixing case sensitive and case insensitive searches by the same column
causes problems on Pg and Oracle -- especially as the case-sensitivity
of the search and any indexes may cause them to not be used.

This adds code to catch, for specific columns, cases where CASESENSITIVE
is not explicitly specified, and throws a warning.