Use File::Temp for non-predictable temporary filenames
Previously, RT wrote to /tmp/rt.form.$$, which is a semi-predictable
filename.  A malicious user could symlink a file of their choosing into
that location, which a privileged user running bin/rt would then
overwrite.  While this race condition would be difficult to exploit, as
it leaves the attacker with no control over the contents of the file,
and relies on beating the race condition, it is still potentially
dangerous.

Use File::Temp to atomically generate a non-predictable filename and
open it for writing.