Tone down the visual prominence of user links in search results
By setting the font-weight and color to inherit, they pick up the
natural look of the search results instead of being bold and the color
depending on visited status.  For aileron, this means the links are
normal weight and always black.  Knowing if you've previously visited a
user summary page is not terribly useful, and the dual-toned links are
distracting in search results.  HTML in formats, such as <b>, can still
modify the links' appearance.