Rename the Configuration menu to Admin and pull it up out of Tools
The admin configuration pages are unlike anything else in the Tools
menu, and the top-level menu will only show up for folks with
ShowConfigTab anyway.  Slightly flattening the menus makes them more
navigable.

The ShowConfigTab right keeps its name since there is no compelling
reason to change it.  Internal consistency is a weak argument since it
is unlikely to confuse us and has had a known meaning for quite some
time.  The benefit of keeping it is users who come across references to
ShowConfigTab in rt-users, the wiki, etc.  When they search for the
right in the UI, they can still find it.  The user-visible description
of the right _is_ updated to reflect the current menu name to provide
confirmation that it is the correct right.