Link the top-level Search menu to the ticket simple search
A functional and reasonable default for a common action like searching
is more useful than a sparse page of duplicate links.

This change increases the visibility of the full simple search page,
which, anecdotally, often seems overlooked.  A quick read of the simple
search explanation immensely increases the usefulness of the
ever-present search box.