Remove filename= suggesions from Content-Disposition lines
No supported standard exists for escaping of the "filename" attribute of
Content-Disposition.  This opens the possibility of various injection
attacks through carefully-crafted filenames if AlwaysDownloadAttachments
is enabled.

As the download endpoint is a dhandler, and most locations already
supply the filename after the final /, browsers will assume the correct
filename even if none is supplied to Content-Disposition.  Hence, remove
the filename= parameter entirely.