Ensure that subjects cannot contain embedded newlines
The default templates embed the ticket's subject in the outgoing Subject
header -- but have no way to ensure that the subject does not contain
newlines, which could be used to insert additional headers.
Programmatically restricting templates from inserting newlines in the
header section is both difficult and overly limiting, as there are valid
reasons to do so.

Strip all newlines from Subject headers.  While this does resolves the
issue for all core RT templates, solving it more generally is
unfortunately not feasable.