Convert all absolute paths in CSS to relative paths…
…thus eliminating the use of Mason substitution for WebPath.

Relative paths are taken all the way to the web root and then back down
from static/ so that the paths work when served in squished form at
NoAuth/css/*/.

If the paths only walked up as high as necessary to descend back down,
they'd break as soon as served from somewhere else.  The paths in this
commit let the CSS be served from anywhere of equal depth.