Differentiate different actions when calling CheckCompatibility
When CheckCompatibility was initially moved out to RT::Handle in
aa9e58d, it flattened the various actions into "pre" and "post" -- and
later 38af184 made all database checks "pre".  As the only test in
CheckCompatibility is comparing to "post", the precise value of "pre" is
irrelevant.

To allow compatibility checks to be precisely targetted at stages of the
rt-setup-database process, pass in a different value for the stage of
each.