Docs: RT->LoadConfig and ->Init need to happen at compile time
Most of the time things will still work without the BEGIN block, but any
plugin that provides a library overlay (such as Ticket_Vendor.pm) won't
correctly load it if the script also contains, continuing the example, a "use
RT::Ticket" line.

This is because use statements are is evaluated at compile-time before
LoadConfig and Init are called, and at that point plugins aren't setup
yet when the core library file looks for overlays to load.