Factor out the common join to the OCFV table
The two calls to _CustomFieldJoin can be moved out and into one call, as
long as care is taken to preserve the logic as to when new indexes are
necessary.  Specifically, new indexes are never necessary with
single-value CFs, and are used if the operator is a negative operator OR
is not /^[<>]=?$/ -- that is, if it is not a relative comparitor.