Stop stripping X-RT-Original-Encoding in GnuPG tests
The function in question is used to take an _outgoing_ message RT sent,
turn it around, clean it up, and reinject it back _into_ RT: the
outgoing GnuPG support is tested with the incoming GnuPG support.

The commit which introduced this code (3a83c99) is huge and offers no
reasoning for stripping the various headers.  It's apparent that some of
the headers need to be stripped from the top-level MIME part so that RT
lets the mail enter RT again without detecting it as a loop.  The regex
approach, however, also removes the listed headers from subparts.  In
particular, X-RT-Original-Encoding is unlike the other headers since
almost all subparts will have it set.  The other headers should only
appear at the top-level.

Modifying the subparts causes decryption/signature verification to fail
since the content no longer matches what was encrypted/signed.  This bug
in the testing infrastructure wasn't noticed with RT's standard plain
text templates since they only have a single textual part.  With the
switch to HTML templates, tests started failing verification.  This
commit makes them pass again.