Preserve original case of defined statuses
When normalizing lifecycle configurations, use the original case that
they were written in for display.  This allows code which depended on
the apparent casing of $ticket->Status to still continue to function,
even while the back-end status is stored in lower-case.

For simplicity of look-up, keys of the lifecycle are stored in
lower-case, while lists of statuses are stored in their canonical case.