When a value overflows into LargeContent, ensure that Content is null
The ->AddCustomFieldValue codepath attempts to ensure that empty values
for Content or LargeContent are never passed down the stack for
insertion.  However, this is foiled if a >255-character value is passed
through for Value, is only caught at the ObjectCustomFieldValue->Create
level -- which sets Content to the empty string when upgrading the
Content to LargeContent, rather than undef.

Ensure that NULLs are properly stored in the Content column when
LargeContent is used, by setting Content to undef in such cases.  There
may also exist historical OCFVs which contain '' instead of NULL; update
them accordingly.