Ensure that system rights do not move to "Status changes" category
RegisterRights checks that a right name is not a pre-existing queue
right before adding it as a new "Status" right.  However, if a right
which is a system right and not a queue right (such as SuperUser) is
used to limit a transition, the category and description are overidden.

When a right name is given, ensure that it is neither a queue right nor
a system right before overriding its category and description.