Replace uses of the now-deprecated LogToScreen with LogToSTDERR
In 9ee4377, the configuration option LogToScreen was renamed to
LogToSTDERR to better describe what it does, and most existing instances
of LogToScreen were renamed.  Recently, 7df7a81 switched to using
RT->Deprecated to warn about it -- and in doing so, dropped the
restriction that the Set( LogToScreen ) must be called from a
configuration file.  This exposed several pieces of code (most of which
were added or merged after 9ee4377) which still attempt to
RT->Config->Set( LogToScreen ), leading to deprecation warnings.

Avoid the deprecation warnings by switching all new uses of LogToScreen
to the correct LogToSTDERR name instead.