Improve != searching in combination with LargeContent
When attempting to perform a != on a single-value CF, with a value
longer than 255 characters, the search is limited by:

  ((Content = '' OR Content IS NULL) AND LargeContent != '...')

The limitation «Content != '...' OR » was intentionally omitted, based
on the proposition that a value longer than 255 characters cannot relate
to the Content field.  However, _any_ value in the Content field
necessarily implies a custom field value which does not match the long
string (because it is too short), meaning that the ticket should be
contained in the resultset.

Conflicts:
	lib/RT/SearchBuilder.pm